backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
celery[redis]==5.4.0
redis==5.2.1
sqlalchemy[asyncio]==2.0.38
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.11.1
pydantic-settings==2.8.1
python-multipart==0.0.20
boto3==1.36.26
pillow==11.1.0
opencv-python-headless==4.11.0.86
scikit-image==0.25.2
numpy==2.2.4
ezdxf==1.4.2
ultralytics==8.3.94
torch==2.6.0
torchvision==0.21.0
segment-anything-2==0.1.0
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.25.3
python-jose[cryptography]==3.3.0
slowapi==0.1.9
sentry-sdk[fastapi]==2.19.2
